Workaround nulls in Proctor API JSON.
Due to a bug in the Proctor REST API, some "include if not null"
annotations are being ignored by Jackson and Spring.

This is a temporary workaround that assumes the API may include
null or multiple nulls in the payload field.

Should be reverted after I fix the problem upstream.