Add method to ProctorGroups with pre-joined groups.
get_group_string() pre-joins the group strings with ','. This
is an incredibly common action for our internal systems, so this
saves using ','.join() everywhere we would use the old method.

Rename get_group_strings() to get_group_string_list(). This
still returns the group strings as a list so that users can
append their own groups before passing this to their logging
infrastructure.